// apps/web-panel-vue/src/views/site/types.d.ts
export interface LetsTableProps { 
  auth_to: string
  auth_type: string
  cert_timeout: number
  certificate_url: string
  domains: string[]
  endDay: number
  expires: number
  index: string
  next_retry_time: number
  renew_time: number
  retry_count: number
  save_path: string
  status: string
}